begins splitting blockbuilder pkgs
